test(mcp-server): broaden tools/list regression to scan full registry
Replace the artifacts-only regression test with one that boots the
canonical registerTools() entry point against a mock server, captures
every tool's inputSchema, and asserts z.toJSONSchema() succeeds on
each. The original bug killed all 33 tools from a single bad schema —
locking only the 6 artifacts tools left the rest of the surface
unprotected against the same bug class.
